Lots more unit testing complete

+++ src/org/interworldtransport/cladosF/ComplexD.java
@@ -523,7 +523,7 @@ public class ComplexD extends UnitAbstract implements Field, Normalizable {
 	 */
 	@Override
 	public Double modulus() {
-		return Double.valueOf((double) Math.sqrt(sqModulus().doubleValue()));
+		return Double.valueOf(Math.sqrt(sqModulus()));
 	}
 
 	/**
